feat(app): allow customizing scorer title, fixed filtering and rescoring issues (#2365)
* fix: fixing issues with filtering, rescoring for customizations

* fix: fixing issues with filtering, rescoring for customizations

* feat(ui): allow customization of scorer title

+++ app/components/PlatformCard.tsx
@@ -141,7 +141,7 @@ const usePlatformIsExcluded = (platform: PlatformScoreSpec) => {
     return (
       isDynamicCustomization(customization) &&
       customization.scorer?.weights &&
-      !providers.some((provider) => parseInt(customization.scorer?.weights?.[provider] || "") > 0)
+      !providers.some((provider) => parseFloat(customization.scorer?.weights?.[provider] || "") > 0)
     );
   }, [customization.key, platform.platform]);
